Get the channel id, root id, and facility id. Moved the getChannelId into core-getters.
